实现 issue #634
