+++ skills/codex-refactor-loop/scripts/test_daemon_heartbeat.py
@@ -8,14 +8,100 @@ import shutil
 import subprocess
 import sys
 import tempfile
+import threading as real_threading
 import unittest
 from pathlib import Path
+from unittest import mock
 
 
 SCRIPT_DIR = Path(__file__).resolve().parent
 sys.path.insert(0, str(SCRIPT_DIR))
 
 from codex_refactor_loop.heartbeat import DaemonHeartbeatLease
+
+REAL_CONDITION = real_threading.Condition
+REAL_EVENT = real_threading.Event
+REAL_THREAD = real_threading.Thread
+
+
+class HeartbeatCoordinator:
+    def __init__(self) -> None:
+        self.condition = REAL_CONDITION()
+        self.callback_entered = False
+        self.beat_during_callback = False
+        self.stop_requested = False
+
+    def enter_callback(self) -> None:
+        with self.condition:
+            self.callback_entered = True
+            self.condition.notify_all()
+
+    def record_beat(self) -> None:
+        with self.condition:
+            if self.callback_entered and not self.stop_requested:
+                self.beat_during_callback = True
+            self.condition.notify_all()
+
+    def wait_for_heartbeat_while_callback_is_pending(self) -> bool:
+        with self.condition:
+            return self.condition.wait_for(lambda: self.beat_during_callback, timeout=1.0)
+
+    def request_stop(self) -> None:
+        with self.condition:
+            self.stop_requested = True
+            self.condition.notify_all()
+
+
+class ScriptedEvent:
+    instances: list["ScriptedEvent"] = []
+    coordinator: HeartbeatCoordinator | None = None
+
+    def __init__(self) -> None:
+        self.wait_timeouts: list[float] = []
+        self.set_called = False
+        ScriptedEvent.instances.append(self)
+
+    def wait(self, timeout: float | None = None) -> bool:
+        self.wait_timeouts.append(float(timeout or 0))
+        if ScriptedEvent.coordinator is None:
+            return True
+        coordinator = ScriptedEvent.coordinator
+        with coordinator.condition:
+            if not coordinator.callback_entered:
+                coordinator.condition.wait_for(lambda: coordinator.callback_entered or coordinator.stop_requested, timeout=1.0)
+            if coordinator.stop_requested:
+                return True
+            if not coordinator.beat_during_callback:
+                return False
+            coordinator.condition.wait_for(lambda: coordinator.stop_requested, timeout=1.0)
+            return True
+
+    def set(self) -> None:
+        self.set_called = True
+        if ScriptedEvent.coordinator is not None:
+            ScriptedEvent.coordinator.request_stop()
+
+
+class InlineThread:
+    instances: list["InlineThread"] = []
+
+    def __init__(self, *, target, name: str, daemon: bool) -> None:
+        self.target = target
+        self.name = name
+        self.daemon = daemon
+        self.started = False
+        with mock.patch("threading.Event", REAL_EVENT):
+            self.thread = REAL_THREAD(target=self.target, name=name, daemon=daemon)
+        self.join_timeouts: list[float] = []
+        InlineThread.instances.append(self)
+
+    def start(self) -> None:
+        self.started = True
+        self.thread.start()
+
+    def join(self, timeout: float | None = None) -> None:
+        self.join_timeouts.append(float(timeout or 0))
+        self.thread.join(timeout=timeout)
 
 
 class DaemonHeartbeatLeaseTests(unittest.TestCase):
@@ -46,6 +132,85 @@ class DaemonHeartbeatLeaseTests(unittest.TestCase):
         self.assertEqual([2.0, 2.0, 1.0], slept)
         self.assertEqual("1003", (self.repo / ".refactor-loop" / "heartbeats" / "python-daemon.ts").read_text().strip())
         self.assertEqual([], list((self.repo / ".refactor-loop" / "heartbeats").glob("*.tmp.*")))
+
+    def test_run_with_lease_periodically_renews_heartbeat_during_callback(self) -> None:
+        ScriptedEvent.instances = []
+        ScriptedEvent.coordinator = HeartbeatCoordinator()
+        InlineThread.instances = []
+        times = iter([2000, 2001])
+        lease = DaemonHeartbeatLease(
+            "python-daemon",
+            self.repo,
+            heartbeat_interval=7,
+            clock=lambda: next(times),
+        )
+        expected = [object()]
+        original_beat = lease.beat
+
+        def beat() -> None:
+            original_beat()
+            if ScriptedEvent.coordinator is not None:
+                ScriptedEvent.coordinator.record_beat()
+
+        lease.beat = beat
+
+        with mock.patch("codex_refactor_loop.heartbeat.threading.Event", ScriptedEvent), mock.patch(
+            "codex_refactor_loop.heartbeat.threading.Thread",
+            InlineThread,
+        ):
+
+            def callback():
+                assert ScriptedEvent.coordinator is not None
+                ScriptedEvent.coordinator.enter_callback()
+                self.assertTrue(ScriptedEvent.coordinator.wait_for_heartbeat_while_callback_is_pending())
+                return expected
+
+            result = lease.run_with_lease(callback)
+
+        self.assertIs(result, expected)
+        self.assertTrue(ScriptedEvent.coordinator.beat_during_callback)
+        self.assertGreaterEqual(ScriptedEvent.instances[0].wait_timeouts.count(7.0), 1)
+        self.assertTrue(ScriptedEvent.instances[0].set_called)
+        self.assertTrue(InlineThread.instances[0].started)
+        self.assertTrue(InlineThread.instances[0].daemon)
+        self.assertEqual(InlineThread.instances[0].name, "python-daemon-heartbeat-renewer")
+        self.assertEqual(InlineThread.instances[0].join_timeouts, [1.0])
+        self.assertEqual("2000", (self.repo / ".refactor-loop" / "heartbeats" / "python-daemon.ts").read_text().strip())
+        ScriptedEvent.coordinator = None
+
+    def test_run_with_lease_propagates_exception_and_stops_renewer(self) -> None:
+        ScriptedEvent.instances = []
+        ScriptedEvent.coordinator = HeartbeatCoordinator()
+        InlineThread.instances = []
+        lease = DaemonHeartbeatLease("python-daemon", self.repo, heartbeat_interval=7, clock=lambda: 2000)
+        original_beat = lease.beat
+
+        def beat() -> None:
+            original_beat()
+            if ScriptedEvent.coordinator is not None:
+                ScriptedEvent.coordinator.record_beat()
+
+        lease.beat = beat
+
+        with mock.patch("codex_refactor_loop.heartbeat.threading.Event", ScriptedEvent), mock.patch(
+            "codex_refactor_loop.heartbeat.threading.Thread",
+            InlineThread,
+        ):
+
+            def callback() -> None:
+                assert ScriptedEvent.coordinator is not None
+                ScriptedEvent.coordinator.enter_callback()
+                self.assertTrue(ScriptedEvent.coordinator.wait_for_heartbeat_while_callback_is_pending())
+                raise RuntimeError("boom")
+
+            with self.assertRaisesRegex(RuntimeError, "boom"):
+                lease.run_with_lease(callback)
+
+        self.assertTrue(ScriptedEvent.coordinator.beat_during_callback)
+        self.assertTrue(ScriptedEvent.instances[0].set_called)
+        self.assertTrue(InlineThread.instances[0].started)
+        self.assertEqual(InlineThread.instances[0].join_timeouts, [1.0])
+        ScriptedEvent.coordinator = None
 
     def test_lease_requires_repo_root_context_or_explicit_heartbeat_file(self) -> None:
         old_env = os.environ.copy()
